remove get*Actions on container
